refactor(cli): split 2987-line cli.rs into focused submodules
Decompose the monolithic src/cli.rs into a src/cli/ directory module,
separating three unrelated concerns that lived in one file:

- args.rs        clap grammar (Commands + 13 subcommand enums, arg
                 structs, From conversions, command policy)
- mod.rs         pub fn run() dispatcher + help helpers
- interactive.rs terminal-interactivity detection
- tests.rs       the existing CLI test module

Behavior-preserving: pure code movement plus pub(crate) visibility on
items now referenced across module boundaries. cli::run() keeps its
signature. No logic changes.

Verified: cargo check/clippy clean (no new warnings), cargo fmt --check
clean, full suite 1699/1699 passing, CLI --help output unchanged.
